Add files via upload
added signing for private users + authenticating for public users. fixed menu so that menu resets and loops properly when user exits the menu they are on. commented a bunch of stuff in explaining what most of the stuff does. There are lots of lists. Don't question the lists. They work. This thing is held together with hopes, dreams, duct tape, and adderall